Sti detectors are not guaranteed to be arranged consecutively
With the addition of iTPC channels in TPC sector 20 the order in which Sti
layers can be arranged by (padrow, sector) is not symmetric anymore, i.e. the
iTPC Sti sector 3 has the padrow with the smallest radius thus taking the (0,3)
slot while, for example, the (0,0) slot has no assigned detector.